Add floating_ip label to node
By adding a label with the floating ip address we can use node selectors
on other pods to assign them to the node.

I have been unable to test the label removal code as I have been unable
to add a node to my cluster which becomes the 0th item in the list
without removing the currently assigned node.